# Conflicts:
#	Paper Soldiers/Assets/Scenes/RuhanScene.unity

+++ Paper Soldiers/Assets/Scripts/Entities/EntityData.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class EntityData : MonoBehaviour
 {
-  public int InitialHealthPoints = 5;
-  public int HealthPoints = 0;
+  public float InitialHealthPoints = 5;
+  public float HealthPoints = 0;
   public float MovementSpeed = 2;
 }
